Split files for PSR-4 compliance
