+++ src/ArrayIterator.php
@@ -1,0 +1,11 @@
+<?php
+namespace JakubOnderka\PhpParallelLint;
+
+class ArrayIterator extends \ArrayIterator
+{
+    public function getNext()
+    {
+        $this->next();
+        return $this->current();
+    }
+}
