+++ src/IWriter.php
@@ -1,0 +1,10 @@
+<?php
+namespace JakubOnderka\PhpParallelLint;
+
+interface IWriter
+{
+    /**
+     * @param string $string
+     */
+    public function write($string);
+}
