+++ src/InvalidArgumentException.php
@@ -1,0 +1,18 @@
+<?php
+namespace JakubOnderka\PhpParallelLint;
+
+class InvalidArgumentException extends Exception
+{
+    protected $argument;
+
+    public function __construct($argument)
+    {
+        $this->argument = $argument;
+        $this->message = "Invalid argument $argument";
+    }
+
+    public function getArgument()
+    {
+        return $this->argument;
+    }
+}
