+++ src/RunTimeException.php
@@ -1,0 +1,7 @@
+<?php
+namespace JakubOnderka\PhpParallelLint;
+
+class RunTimeException extends Exception
+{
+
+}
